First step in modularizing Kafka-Monitor
Renamed settings_crawling -> scraper_settings

This is the first step at modularizing the validation and handling of incoming requests. Each handler will have a schema and a 'successful' processing method, which kafka-monitor.py will remain agnostic. This is useful as we continue to expand and need more styles of validation, it makes no sense to bloat the core kafka-monitor code.

Moved some code into the handlers/ directory too.