Added logging to crawler
Added a bunch of new logging to the crawler, specifically to help debug efforts or to see how logic flows. In combination with the schema commit prior, removed a bunch of null field setups in the `next_request()` method to help clean that up a bit so it was not soley dependent on our current schema.