Changed default_settings.py to settings.py to better conform to Scrapy standards
Every different setting you want locally should be in a file named `localsettings.py` and will a) override the defaults and b) not be tracked by git.

The behavior for overriding defaults is not quite the same in the kafka/redis monitor vs scrapy, but the difference is pretty small.