Added Redis Monitor stats collection
Added stat collection for plugins and total successes/failures. Needed to add a precondition check for the redis monitor plugins because there may be redis keys that are not actually acted on, and therefore should not be used to increment stats. The first use case is the Expire Monitor, and the second use case will be a new plugin used to log all the stats out (if they exist).

Also added two new unit tests to ensure the functionality is captured.