Plugins no longer responsible for deleting keys.
This was the original design and I am reverting back to it, Redis Monitor plugins should not be responsible for deleting the key they are acting on. If an exception is thrown in the plugin, the delete key code is never reached create an infinte loop of errors.

This new/original way allows for consise logs and for the plugin to just do it's processing on the key and move on.