Added distributed locks to redis monitor for scaling
This commit adds the ability for the Redis Monitor to scale out horizontally to N number of processes. Each process uses a non-blocking lock to find and retrieve a job or action to then operate on. Most of the time the job completes very quickly, but with extremely large keystores things may take longer.

While executing and action or job, the redis monitor is unavailable to be tasked again since the job is done on the main thread and is blocking. If you would like to execute more than one action at once, you can scale the number of redis monitor processes running to support your needs. Each action is only worked on by one process at a time thanks to the lock, so you wont get unexpected duplicate data.

The actions/jobs also have a retry failover, in case the process is killed or exits unexpectedly.

*Overall, this eliminates the single point of failure that is the Redis Monitor.* All three components are now able to be scaled to the application's desire.

Updates include:
- Documentation surrounding new settings parameters
- Locking implementation
- Updated unit tests for new components
- One new package

fixes #43