Add Docker Support (#72)
* Beginning work on docker branch.

Work here will focus on creating Dockerfiles for the crawler, kafka monitor, and redis monitor. Will use supervisord as a base with environment variables to alter/update any kind of configuration we need within either the python settings.py files or the supervisor configs.

references #48 #26

* Kafka Monitor Docker

Adds a dockerfile to setup and run the kafka monitor. I do note that there are settings and configuration within that tie it into the docker-compose file, but I think they can be fairly easily overridden and maybe I switch them to just using the raw `settings.py` file instead of `localsettings.py`.

This uses stand alone containers for Kafka, Zookeeper, and Redis, and each of the smaller scrapy cluster components
will be their own separate containers as well.

Everything will run under supervisord so you can scale the number of processes both within the container and outside. There is much more work to be done, for example reading in environment variables to override things in the settings file, configuring docker hub, etc.

* Added Redis Monitor to docker setup.

This is an initial cut at making the redis monitor docker compatible. There are still many environmen
t variables to specify but this gets us prett far along

In testing this I also found a dormant issue with ensuring the ZK file path existed, which is easy enough to fix and will be merged in when this branch is decently complete

* Added Dockerfile for crawler

This commit adds the link spider to be compatible with Docker. This completes the initial cut of dockerization of the three core components, and todo is the following:

- Define environment variable overrides for commonly altered configurations
- Documentation under the Advanced Topics as well as Quickstart to supply yet another alternative to working and provisioning Scrapy Cluster.
- Update Changelog
- Merge branch back into dev.
- Dockerhub builds, or at least a stable set of images on the hub for people to pull down
- ???

Very excited to get this going!

* Added commonly altered variables to the settings.py files

* Removed supervisord, Docker can restart the containers and scale

* prep for dev vs prod docker compose

* Added docker-compose based off of dockerhub images

* Add initial docker docs

* Updated testing for docker

This commit hopefully enables us to run the integration tests for our Docker images within Travis, enabling us to continuously test both the ansible provisioning and docker images.

Added a new script to run tests within the container. Updated documentation for quickstart and advanced docker docs. Modded the Dockerfiles to include the new test script. Fixed .gitignore to ignore more files

* Travis build matrix testing

Trying to make life easier by separating things into shell scripts instead of a massive script section in travis

* more travis changes

* more travis configs

* fix variable reference

Still need to figure out why kafka doesnt start immediately

* More tweaks to get kafka and kafka-monitor to stand up first try

* rename test script folder to just "travis"

* trying to get kafka to spin up 1st try in travis

* forgot to fix .test ports

* kafka monitor typo

* Add docker hub pulls badge

* move travis test compose file into travis folder

* travis name change with file move

* docker down correct file path in travis

* clean up docs, remove old supervisord.conf