Begin work on rest endpoint
This branch is beginning the work on the rest endpoint outlined at #24. This adds a new flask rest endpoint that will be used to work with the scraping cluster, and interact with the kafka api.

The work will reside in `/rest` and will eventually have both offline test, online tests, and a new section of documentation. Notable new files here:

* rest_service.py - the rest endpoint code
* settings.py - rest service settings
* requirements.txt - the bare minimum requirements to run the rest service