Updated rest service for kafka failsafe mechanisms
Adds the ability for both the consumer and the producer to detect (to the best of their ability) when kafka goes down, and respawn the kafka setup thread. Also uses a thread lock to make updates to the dictionary used to pass the response messages from the daemon consumer thread.

Adds a couple new settings as well.