Added poll endpoint for requests that take longer than the threshold
This commit adds a new `/poll` endpoint that accepts a `poll_id` returned from `/feed` and allows you
 to query for it at a later time. This is useful for **stop** requests or any other request from the
redis monitor that takes longer than the default timeout.

Since we use redis to cache the result, you can even put the rest service behind a load balancer and
not have to worry about which application actually made the request - when it is finished it will be
available to the other nodes as well.

Also adds a schema validator for future endpoints so we can be sure that the data coming into the res
t endpoint is exactly like we expect.