Add kafka-python future callbacks
This commit adds future callbacks everywhere we send data to Kafka. This helps us better track the data is actually making it to kafka, where before we only assumed. Updates
* Rest Service
* Redis Monitor
* Crawler

This removes the need for a second validation step within the Scrapy Item Pipeline, and updates the kafka-python v 1.3.2.

Scrapy 1.3.1 update as well.

Closes #93