Purge Ubuntu 16.04 from ansible
Everything I have tried requires systemctl daemon reload, which continues to cause troubles when ran on Travis inside of a Docker container. Reverting back so this is not a huge time sink.

To be clear, 16.04 is supported, however the way the ansible scripts are set up to run kafka, zookeeper, and redis are not.