Update scutils to 1.3.0dev0
Adds python 2/3 compatibility for LogFactory. Also updates how we build the `scutils` library inside the project to allow for proper pull requests against scutils to be built and tested by travis.

This is done via a special flag in the requirements to point to a folder rather than a pypi package.

This is the first cut so I expect some errors to occur in the build process