Initial Support for Splash:
Lots of changes here in order to get things working. The scrapy-splash middleware generates a new request that must be handled by the scheduler, so we need additional logic in order to both put the request back into the queue, but to also pop it off and work with it.

To use splash, add `"splash":{}` to your kafka monitor request. That will tell scrapy cluster to use splash to fetch the webpage in question. I tried to leave as many options as possible for the end user, however I don't know all the different use cases and I know there are some hard coded values here to render the html of the page.

You may populate the splash object in the request with anything you will need for that initial request. Keep in mind that it will be propigated if `maxdepth>0`, so additional logic or a custom spider may be needed for special processing or other logic.

Right now, my concern is not for supporting special styles of Splash requests, as I think that is better suited for normal Scrapy operations. A form request might not come back to the same spider, or same ip, so given that the whole purpose is to be stateless and distributed I think custom setups can serve that better than a generic solution.

There are other minor changes throughout the crawler codebase to support `maxdepth>0` requests, like the meta passthrough middleware updates or the log retry middleware update. This also updates the DuplicationFilter based on the Scrapy Splash dupefilter logic.

This should work out of the box with Docker and Docker Compose, however more testing/validation is needed.

Todo:
- Unit Tests
- Integration Tests for Splash
- Ansible setup to test splash on ubuntu/centos
- Documentation
- Further Validation
- Traefik integration (for scaling splash processes behind a proxy)

Other notes:
- Because Splash Requests are separate, this causes 2x the number of requests, and thus the crawl rate is 2x slower when generating requests that include splash. One time to get the normal request, then again to get the newly created splash request.

Relates to: #108